requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofile==3.8.8
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.0.6
anyio==3.6.2
anykeystore==0.2
apex==0.1
appdirs==1.4.4
apt-xapian-index==0.49
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==21.2.0
azure-core==1.29.5
azure-identity==1.15.0
azure-quantum==0.29.2
azure-storage-blob==12.19.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
bleach==6.0.0
blinker==1.4
blobfile==2.0.0
bokeh==3.0.3
cached-property==1.5.2
cachetools==5.3.2
caio==0.9.13
celery==5.1.2
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cluster @ git+https://code.it4i.cz/def/cluster.git@428bd7f6c15525942358f5ff6b54d07b3828025f
codecov==2.1.12
colorama==0.4.3
coloredlogs==15.0.1
comm==0.1.2
command-not-found==0.3
commonmark==0.9.1
ConfigArgParse==1.5.3
connection-pool==0.0.3
contourpy==1.0.7
coverage==6.5.0
cryptacular==1.6.2
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.24
dacite==1.8.0
dataclasses==0.6
datrie==0.8.2
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distlib==0.3.3
distro==1.4.0
distro-info==0.23+ubuntu1.1
docker-pycreds==0.4.0
docutils==0.17.1
entrypoints==0.3
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.3.0
flake8==5.0.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.7
GitPython==3.1.24
google-auth==2.23.4
google-auth-oauthlib==1.0.0
gpg==1.13.1
greenlet==1.1.1
grpcio==1.59.2
httplib2==0.14.0
huggingface-hub==0.14.1
humanfriendly==10.0
hupper==1.10.3
idna==2.8
imageio==2.9.0
importlib-metadata==5.0.0
importlib-resources==5.10.2
-e git+https://github.com/User3574/improved-diffusion@e168fcdfe0bc8f7ac237000c93d472a2d799e609#egg=improved_diffusion
iniconfig==1.1.1
interrogate==1.5.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
isort==4.3.21
iunets==0.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==8.0.2
jupyter-console==6.6.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.2.1
jupyter-server-terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keyring==18.0.1
kiwisolver==1.3.2
kombu==5.1.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.1
maestrowf==1.1.7.dev0
Markdown==3.4.1
MarkupSafe==2.1.2
mashumaro==2.9
matplotlib==3.4.3
matplotlib-inline==0.1.6
mccabe==0.7.0
MechanicalSoup==1.1.0
memcnn==1.5.2
merlin==1.8.1
mistune==2.0.5
mmcls==0.24.0
mmcv-full==1.6.2
-e git+https://github.com/open-mmlab/mmgeneration.git@f6551e1d6ca24121d1f0a954c3b3ac15de6d302e#egg=mmgen
model-index==0.1.11
mpi4py==3.1.3
mpmath==1.3.0
msal==1.26.0
msal-extensions==1.0.0
msgpack==1.0.3
msrest==0.7.1
multidict==6.0.4
multiprocess==0.70.14
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.10.4
networkx==2.8.7
ninja==1.10.2.4
notebook==6.5.2
notebook-shim==0.2.2
ntlm-auth==1.5.0
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
openmim==0.3.2
ordered-set==4.1.0
packaging==20.9
pandas==1.3.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
PasteDeploy==2.1.1
pathlib2==2.3.7.post1
pathtools==0.1.2
pbkdf2==1.3
pbr==6.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
plaster==1.0
plaster-pastedeploy==0.7
platformdirs==3.2.0
pluggy==0.13.1
ply==3.11
portalocker==2.8.2
prettytable==3.4.1
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==4.25.0
psutil==5.9.4
ptyprocess==0.7.0
PuLP==2.5.1
pure-eval==0.2.2
py==1.10.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycodestyle==2.9.1
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.15.0
pycups==1.9.73
pyflakes==2.5.0
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==1.7.1
pylatexenc==2.10
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyprof2calltree==1.4.4
pyqir==0.10.0
PyQt5==5.14.1
pyramid==2.0
pyramid-mailer==0.15.1
pyrsistent==0.18.0
pytest==6.2.4
pytest-runner==6.0.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-json-logger==2.0.4
python-markdown-math==0.8
python3-openid==3.2.0
pytz==2021.1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.0.0
qiskit==0.43.3
qiskit-aer==0.12.2
qiskit-ibmq-provider==0.20.2
qiskit-ionq==0.4.4
qiskit-qir==0.4.0
qiskit-terra==0.24.2
qtconsole==5.4.1
QtPy==2.3.1
ratelimiter==1.2.0.post0
recordclass==0.16.1
redis==3.5.3
regex==2023.10.3
reportlab==3.5.34
repoze.sendmail==4.4.1
requests==2.31.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
responses==0.18.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
rustworkx==0.13.2
scikit-image==0.19.3
scikit-learn==0.24.2
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.8.0
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
SimpleITK==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0
smart-open==5.2.1
smmap==4.0.0
snakemake==6.9.1
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==1.4.23
ssh-import-id==5.10
stack-data==0.6.2
stevedore==5.1.0
stopit==1.1.2
symengine==0.9.2
sympy==1.12
systemd-python==234
tabulate==0.8.9
tensorboard==2.14.0
tensorboard-data-server==0.7.2
terminado==0.17.1
threadpoolctl==2.2.0
tifffile==2022.8.12
tinycss2==1.2.1
tk==0.1.0
toml==0.10.2
toposort==1.7
torch==1.12.1
torchvision==0.13.1
tornado==6.2
tqdm==4.62.2
traitlets==5.9.0
transaction==3.0.1
translationstring==1.4
typer==0.4.0
typing-extensions==4.8.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uri-template==1.2.0
urllib3==1.25.8
usb-creator==0.3.7
velruse==1.1.1
venusian==3.0.0
vine==5.0.0
virtualenv==20.8.1
wadllib==1.3.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.5.1
websockets==12.0
werkzeug==3.0.1
widgetsnbextension==4.0.7
wrapt==1.13.1
WTForms==2.3.3
wtforms-recaptcha==0.3.2
xkit==0.0.0
xxhash==3.2.0
xyzservices==2023.2.0
yacs==0.1.8
yapf==0.32.0
yarl==1.9.2
zipp==3.9.0
zmq==0.0.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
